tambunan

+++ app/src/main/java/com/somboi/rodaimpian/gdx/entities/WheelParam.java
@@ -4,5 +4,5 @@ public class WheelParam {
     public int resultValue;
     public String results;
     public float wheelangle;
-    public int wheelImpulse;
+    public float wheelImpulse;
 }
